% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetry.R
\name{calculate_asymmetry}
\alias{calculate_asymmetry}
\title{Calculate asymmetry values}
\usage{
calculate_asymmetry(input_data, threshold = NULL, TI_steps = 3, filter = TRUE)
}
\arguments{
\item{input_data}{Input igraph object or data frame}

\item{TI_steps}{TI steps to compute}

\item{filter}{Threshold to use, default is top 1\% of asymmetric links}
}
\value{
A data frame with links indicating direction of asymmetric effect and it's value.
}
\description{
This function calculates interaction asymmetry based on TI index for networks.
Input should be an igraph object or a data frame without header,
the columns should be ordered: node name of species A,
node name of species B, weight (if exists).
If filter == TRUE (default) then top 1\% of asymmetric links will be returned. Other absolute
threshold values can be supplied.
}
