% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TI_WI.R
\name{calculate_TI_WI}
\alias{calculate_TI_WI}
\title{Calculate TI/WI index}
\usage{
calculate_TI_WI(
  input_data,
  steps,
  symmetrization_method = "sum",
  asymmetry = F
)
}
\arguments{
\item{input_data}{Input igraph object or data frame}

\item{steps}{TI steps to compute}

\item{symmetrization_method}{How to combine TI values}

\item{asymmetry}{Whether to return an input for asymmetry calculation}
}
\value{
A data frame with node ids and TI, WI (if there are weights) index.
}
\description{
This function calculates the TI and WI index for networks. Input should be an igraph object or
a data frame without header, the columns should be ordered: node name of species A,
node name of species B, weight (if exists).
Symmetrization_method is optional, can be either 'sum' (default), 'average', 'max', 'min' or
'difference.' Multilink edges will be summed.
}
